third_party/WebKit/LayoutTests/dom/document_fragment/parent-node-interface-expected.txt

This tests that DocumentFragment implements the ParentNode interface.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS documentFragment.children.length is 0
PASS documentFragment.childElementCount is 0
PASS documentFragment.firstElementChild is null
PASS documentFragment.lastElementChild is null
PASS documentFragment.children.length is 0
PASS documentFragment.childElementCount is 0
PASS documentFragment.firstElementChild is null
PASS documentFragment.lastElementChild is null
PASS documentFragment.children.length is 1
PASS documentFragment.childElementCount is 1
PASS documentFragment.children[0] is b
PASS documentFragment.firstElementChild is b
PASS documentFragment.lastElementChild is b
PASS documentFragment.children.length is 1
PASS documentFragment.childElementCount is 1
PASS documentFragment.children[0] is b
PASS documentFragment.firstElementChild is b
PASS documentFragment.lastElementChild is b
PASS documentFragment.children.length is 2
PASS documentFragment.childElementCount is 2
PASS documentFragment.children[0] is b
PASS documentFragment.children[1] is d
PASS documentFragment.firstElementChild is b
PASS documentFragment.lastElementChild is d
PASS documentFragment.children.length is 2
PASS documentFragment.childElementCount is 2
PASS documentFragment.children[0] is b
PASS documentFragment.children[1] is d
PASS documentFragment.firstElementChild is b
PASS documentFragment.lastElementChild is d
PASS successfullyParsed is true

TEST COMPLETE
